Update budget.py
Aha! Finally you can make a budget with this code, and if you wanted as well you can use the create_spend_chart() to get a bar graph showing you the percentage you've spent. 
This code still needs better commenting and cleaner code. That's the next thing I'll be working on for all my projects.